+++ src/Symfony/Bundle/SecurityBundle/Resources/config/security_authenticator.php
@@ -25,6 +25,7 @@ use Symfony\Component\Security\Http\EventListener\AuthenticationProofsListener;
 use Symfony\Component\Security\Http\EventListener\CheckCredentialsListener;
 use Symfony\Component\Security\Http\EventListener\LoginThrottlingListener;
 use Symfony\Component\Security\Http\EventListener\PasswordMigratingListener;
+use Symfony\Component\Security\Http\EventListener\RefreshedUserCheckerListener;
 use Symfony\Component\Security\Http\EventListener\SessionStrategyListener;
 use Symfony\Component\Security\Http\EventListener\UserCheckerListener;
 use Symfony\Component\Security\Http\EventListener\UserProviderListener;
@@ -90,6 +91,12 @@ return static function (ContainerConfigurator $container) {
             ->tag('kernel.event_subscriber')
 
         ->set('security.listener.user_checker', UserCheckerListener::class)
+            ->abstract()
+            ->args([
+                abstract_arg('user checker'),
+            ])
+
+        ->set('security.listener.user_checker_on_refresh', RefreshedUserCheckerListener::class)
             ->abstract()
             ->args([
                 abstract_arg('user checker'),